+++ False_Prophet/js/states/Second_Level.js
@@ -87,23 +87,6 @@ Second_Level.prototype = {
 		overlay = game.add.image(0, 0, 'overlay');
 		overlay.scale.x = .7;
 		overlay.scale.y = .7;
-
-		//adds charts to the overlay for reference
-		var chart = game.add.image(0, 0, 'spritesheet', 'disguisechart');
-		chart.anchor.set(.5);
-		chart.scale.x = .25;
-		chart.scale.y = .25;
-		chart.fixedToCamera = true;
-		chart.cameraOffset.setTo(45, game.height - 23);
-		game.world.bringToTop(chart);
-		chart = game.add.image(0, 0, 'spritesheet', 'trianglechart');
-		chart.anchor.set(.5);
-		chart.scale.x = .25;
-		chart.scale.y = .25;
-		chart.fixedToCamera = true;
-		chart.cameraOffset.setTo(660, game.height - 35);
-		game.world.bringToTop(chart);
-
 		//fixes the overlay to the camera
 		overlay.fixedToCamera = true;
 
@@ -214,6 +197,22 @@ Second_Level.prototype = {
 		levelcomplete.anchor.setTo(.5);
 		cooldown.fixedToCamera = true;
 		levelcomplete.fixedToCamera = true;
+		game.world.bringToTop(overlay);
+		//adds charts to the overlay for easier player reference
+		var chart = game.add.image(0, 0, 'spritesheet', 'disguisechart');
+		chart.anchor.set(.5);
+		chart.scale.x = .25;
+		chart.scale.y = .25;
+		chart.fixedToCamera = true;
+		chart.cameraOffset.setTo(45, game.height - 23);
+		game.world.bringToTop(chart);
+		chart = game.add.image(0, 0, 'trianglechart');
+		chart.anchor.set(.5);
+		chart.scale.x = .35;
+		chart.scale.y = .35;
+		chart.fixedToCamera = true;
+		chart.cameraOffset.setTo(645, game.height - 45);
+		game.world.bringToTop(chart);
 
 
 	},
@@ -248,7 +247,7 @@ Second_Level.prototype = {
 		
 		//not a great solution, but if player is always set to the collide, changing the body type and shape won't phase it
 		player.body.setCollisionGroup(playerCollisionGroup);
-		game.world.bringToTop(overlay);
+		
 
 		//brings the cooldown text to the top of the screen and moves it
 	
@@ -257,8 +256,8 @@ Second_Level.prototype = {
     	levelcomplete.cameraOffset.setTo(game.width/2, 200);
 		game.world.bringToTop(levelcomplete);
 		//position chart HUD
-		chart.cameraOffset.setTo(45,game.height-23);
-		game.world.bringToTop(chart);
+		//chart.cameraOffset.setTo(45,game.height-23);
+		//game.world.bringToTop(chart);
 
 		//changes the text of the cooldown dependant on the player's state
 		if (player.cooldownLeft < .1){
@@ -487,14 +486,16 @@ Second_Level.prototype = {
 	
 	killPlayer: function(playershape, enemyshape){
 		//kills the player when collided with any shape
-		this.createParticles(player, false);
-		player.animations.stop();
-		var rng = Math.floor(Math.random()*3);
-		poofArray[rng].play(null, 0, 1, false, false);
-		player.destroyed = true;
-		player.kill();
-		//restarts the level after we see the particle explosion of the player
-		game.time.events.add(Phaser.Timer.SECOND * 2, function() { game.state.start('Second_Level')});
+		if (!player.destroyed){
+			this.createParticles(player, false);
+			player.animations.stop();
+			var rng = Math.floor(Math.random()*3);
+			poofArray[rng].play(null, 0, 1, false, false);
+			player.destroyed = true;
+			player.kill();
+			//restarts the level after we see the particle explosion of the player
+			game.time.events.add(Phaser.Timer.SECOND * 2, function() { game.state.start('Second_Level')});
+		}
 	},
 	nextLevel: function(){
 		//for when the player runs into the ending block
